Revert "use uint32_t for iterators over toxes in conference_test"
This reverts commit d328b045a36ae19303530e96ceaba0a3760719f3.

+++ auto_tests/conference_test.c
@@ -78,7 +78,7 @@ static void handle_conference_message(
 
 static void run_conference_tests(Tox **toxes, uint32_t *tox_index)
 {
-    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
         tox_callback_conference_message(toxes[i], &handle_conference_message);
     }
 
@@ -92,7 +92,7 @@ static void run_conference_tests(Tox **toxes, uint32_t *tox_index)
     num_recv = 0;
 
     for (uint8_t j = 0; j < 20; ++j) {
-        for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+        for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
             tox_iterate(toxes[i], &tox_index[i]);
         }
 
@@ -102,8 +102,8 @@ static void run_conference_tests(Tox **toxes, uint32_t *tox_index)
     c_sleep(25);
     ck_assert_msg(num_recv == NUM_GROUP_TOX, "failed to recv group messages");
 
-    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
-        for (uint32_t j = 0; j < NUM_GROUP_TOX; ++j) {
+    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
+        for (uint16_t j = 0; j < NUM_GROUP_TOX; ++j) {
             const int len = tox_conference_peer_get_name_size(toxes[i], 0, j, nullptr);
             ck_assert_msg(len == NAMELEN, "name of #%d according to #%d has incorrect length %d", tox_index[j], tox_index[i], len);
             uint8_t name[NAMELEN];
@@ -116,18 +116,18 @@ static void run_conference_tests(Tox **toxes, uint32_t *tox_index)
         }
     }
 
-    for (uint32_t k = NUM_GROUP_TOX; k != 0 ; --k) {
+    for (uint16_t k = NUM_GROUP_TOX; k != 0 ; --k) {
         tox_conference_delete(toxes[k - 1], 0, nullptr);
 
         for (uint8_t j = 0; j < 10; ++j) {
-            for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+            for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
                 tox_iterate(toxes[i], &tox_index[i]);
             }
 
             c_sleep(50);
         }
 
-        for (uint32_t i = 0; i < k - 1; ++i) {
+        for (uint16_t i = 0; i < k - 1; ++i) {
             uint32_t peer_count = tox_conference_peer_count(toxes[i], 0, nullptr);
             ck_assert_msg(peer_count == (k - 1), "\n\tBad number of group peers (post check)."
                           "\n\t\t\tExpected: %u but tox_instance(%u) only has: %u\n\n",
@@ -149,7 +149,7 @@ static void test_many_group(void)
 
     printf("creating %d toxes\n", NUM_GROUP_TOX);
 
-    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
         TOX_ERR_NEW err;
         tox_index[i] = i + 1;
         toxes[i] = tox_new_log(opts, &err, &tox_index[i]);
@@ -195,7 +195,7 @@ static void test_many_group(void)
     while (online_count != NUM_GROUP_TOX) {
         online_count = 0;
 
-        for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+        for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
             tox_iterate(toxes[i], &tox_index[i]);
             online_count += tox_friend_get_connection_status(toxes[i], 0, nullptr) != TOX_CONNECTION_NONE;
         }
@@ -215,10 +215,10 @@ static void test_many_group(void)
                   "failed to set group title");
 
 
-    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
         if (tox_self_get_friend_list_size(toxes[i]) > 1) {
             while (tox_conference_invite(toxes[i], 1, 0, nullptr) == 0) {
-                for (uint32_t j = 0; j < NUM_GROUP_TOX; ++j) {
+                for (uint16_t j = 0; j < NUM_GROUP_TOX; ++j) {
                     tox_iterate(toxes[j], &tox_index[j]);
                 }
             }
@@ -237,7 +237,7 @@ static void test_many_group(void)
         invited_count = 0;
         printf("current peer counts: [");
 
-        for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+        for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
             tox_iterate(toxes[i], &tox_index[i]);
             uint32_t peer_count = tox_conference_peer_count(toxes[i], 0, nullptr);
             invited_count += peer_count == NUM_GROUP_TOX;
@@ -255,7 +255,7 @@ static void test_many_group(void)
         c_sleep(200);
     }
 
-    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
         uint32_t peer_count = tox_conference_peer_count(toxes[i], 0, nullptr);
 
         ck_assert_msg(peer_count == NUM_GROUP_TOX, "\n\tBad number of group peers (pre check)."
@@ -275,7 +275,7 @@ static void test_many_group(void)
 
     printf("tearing down toxes\n");
 
-    for (uint32_t i = 0; i < NUM_GROUP_TOX; ++i) {
+    for (uint16_t i = 0; i < NUM_GROUP_TOX; ++i) {
         tox_kill(toxes[i]);
     }
